interfaz mozo terminada a medias

+++ src/app/pages/home/waiter-home/waiter-home.page.ts
@@ -1,6 +1,10 @@
 import { Component, OnInit } from '@angular/core';
+import { elementAt } from 'rxjs';
 import { enumCollectionNames } from 'src/app/enums/collectionNames';
+import { orderState } from 'src/app/enums/orderState';
+import { order } from 'src/app/interfaces/order';
 import { DataBaseService } from 'src/app/services/data-base.service';
+import { UtilsService } from 'src/app/services/utils.service';
 
 @Component({
   selector: 'app-waiter-home',
@@ -11,33 +15,70 @@ export class WaiterHomePage implements OnInit
 {
   public arrayPedidos : Array<any> = [];
   public arrayPagos : Array<any> = [];
-  constructor(private firebase: DataBaseService) { }
+  constructor(private firebase: DataBaseService, private util : UtilsService) { }
 
   ngOnInit() 
   {
     this.firebase.getObservable(enumCollectionNames.Orders).subscribe((ordenes)=>{
+      this.arrayPedidos = [];
       this.arrayPedidos = [...ordenes];
       this.arrayPedidos.forEach((pedido)=>{
-        if(pedido.cocinaTerminado && pedido.barTerminado)
+        if(pedido.state == orderState.InPreparation && (pedido.kitchenFinished && pedido.barFinished))
           {
-            pedido.estadoPedido = "terminado";
+            pedido.state = orderState.Finished;
           }
       })
+      this.util.SendPushNotification("Nuevo Pedido", "Una mesa ah hecho un nuevo pedido");
     })
   }
 
-  Aceptar(pedido : any)
-  {
-    //verificar tipo de producto y cmabiar el estado de cocina y/o bartender
+  Aceptar(pedido : order)
+  { 
+    this.util.showSweet({title: "Pedido", text: "Estas seguro de confirmar el pedido", showCancelButton: true,
+      cancelButtonText:"No, cambie de opinión", confirmButtonText: "Estoy seguro" , cancelButtonColor: "red"})
+    .then((result)=>
+    {
+      if(result.isConfirmed)
+      {
+        pedido.products.find((element)=>{
+          if(element.isFood)
+          {
+            pedido.kitchenFinished = false;
+          }
+        })
+        pedido.products.find((element)=>{
+          if(!element.isFood)
+          {
+            pedido.barFinished = false;
+          }
+        })
+        pedido.state = orderState.InPreparation;
+        this.firebase.saveData(enumCollectionNames.Orders,pedido,pedido.id)
+      }
+    })
   }
 
-  Eliminar(pedido : any)
+  Eliminar(pedido : order)
   {
-    this.firebase.deleteData(enumCollectionNames.Orders,pedido.id);
+    this.util.showSweet({title: "Pedido", text: "Estas seguro de eliminar el pedido", showCancelButton: true,
+      cancelButtonText:"No, cambie de opinión", confirmButtonText: "Estoy seguro" , cancelButtonColor: "red"})
+    .then((result)=>
+    {
+      if(result.isConfirmed)
+      {
+        this.firebase.deleteData(enumCollectionNames.Orders,pedido.id);
+      }
+    })
   }
 
-  Llevar(pedido : any)
+  Llevar(pedido : order)
   {
-    //cambiar estado pedido a llevar
+    pedido.state = orderState.GivingOut;
+    this.firebase.saveData(enumCollectionNames.Orders,pedido,pedido.id)
   }
+
+  /*AceptarPaga(paga : order)
+  {
+
+  }*/
 }
